# backend-ai-engineering/w2-a1-build-your-first-CRUD-API/js-track/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "A simple CRUD API for tasks"
  },
  "servers": [{ "url": "http://localhost:3000" }],
  "paths": {
    "/": {
      "get": {
        "summary": "API info",
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/tasks": {
      "get": {
        "summary": "List tasks",
        "responses": {
          "200": {
            "description": "A list of tasks",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TasksResponse" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NewTask" }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/tasks/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "integer" }
        }
      ],
      "get": {
        "summary": "Get task by id",
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Not Found" }
        }
      },
      "put": {
        "summary": "Update task",
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Not Found" }
        }
      },
      "delete": {
        "summary": "Delete task",
        "responses": {
          "204": { "description": "No Content" },
          "404": { "description": "Not Found" }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": { "200": { "description": "OK" } }
      }
    }
  },
  "components": {
    "schemas": {
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "done": { "type": "boolean" }
        }
      },
      "TasksResponse": {
        "type": "object",
        "properties": {
          "tasks": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Task" }
          }
        }
      },
      "NewTask": {
        "type": "object",
        "required": ["title"],
        "properties": { "title": { "type": "string" } }
      }
    }
  }
}